FastAPI/requirements.txt
# Core libraries for evaluation and serving
pandas==1.5.0
scikit-learn==1.5.2
numpy==1.26.4
scipy
mlflow-skinny  # Lightweight client for model loading
nltk==3.8.1
boto3          # AWS SDK for Python, needed by MLflow for S3 access

# Libraries for FastAPI & Prometheus
fastapi
uvicorn[standard]
prometheus-client
prometheus-fastapi-instrumentator
jinja2
python-multipart

# Library for making async HTTP requests
aiohttp

# Timezone handling
pytz
